test: 인증 없는 메모 실패 테스트
